% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{build_generic_obj_plot}
\alias{build_generic_obj_plot}
\title{Build diagnostic scatter plots for a computed objective}
\usage{
build_generic_obj_plot(merged_data, value_map_dt, plot_labels)
}
\description{
Build diagnostic scatter plots for a computed objective
}
\keyword{internal}
