% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_design.R
\name{.archived_filename}
\alias{.archived_filename}
\title{Archived filename: \verb{run_<id>_<original basename>}}
\usage{
.archived_filename(run_id, path)
}
\description{
Archived filename: \verb{run_<id>_<original basename>}
}
\keyword{internal}
