% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_design.R
\name{.as_param_design}
\alias{.as_param_design}
\title{Attach metadata used by \code{\link[=extend_param_design]{extend_param_design()}} / \code{\link[=run_param_design]{run_param_design()}}}
\usage{
.as_param_design(df, bounds, method, seed, n_drawn)
}
\description{
Attach metadata used by \code{\link[=extend_param_design]{extend_param_design()}} / \code{\link[=run_param_design]{run_param_design()}}
}
\keyword{internal}
