% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_design.R
\name{.check_unique_basenames}
\alias{.check_unique_basenames}
\title{Ensure selected outputs don't share a basename (they would collide when flattened)}
\usage{
.check_unique_basenames(output_files)
}
\description{
Ensure selected outputs don't share a basename (they would collide when flattened)
}
\keyword{internal}
