% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_design.R
\name{.copy_design_outputs}
\alias{.copy_design_outputs}
\title{Copy selected Daisy output files into keep_dir as run_\if{html}{\out{<id>}}_\if{html}{\out{<filename>}}}
\usage{
.copy_design_outputs(output_files, src_root, keep_dir, run_id)
}
\description{
Copy selected Daisy output files into keep_dir as run_\if{html}{\out{<id>}}_\if{html}{\out{<filename>}}
}
\keyword{internal}
