% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dai_management.R
\name{.dai_parse_harvest_like}
\alias{.dai_parse_harvest_like}
\title{Shared by \code{harvest} and \code{cut}'s inverse.}
\usage{
.dai_parse_harvest_like(ch)
}
\description{
Shared by \code{harvest} and \code{cut}'s inverse.
}
\keyword{internal}
