% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{.dds_optimize}
\alias{.dds_optimize}
\title{Dynamically Dimensioned Search (Tolson & Shoemaker 2007)}
\usage{
.dds_optimize(fn, par, lower, upper, control = list(), ...)
}
\description{
Dynamically Dimensioned Search (Tolson & Shoemaker 2007)
}
\keyword{internal}
