% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_config.R
\name{.default_plf_curve_params}
\alias{.default_plf_curve_params}
\title{Example shape parameters used by \code{\link[=plot_plf_curves]{plot_plf_curves()}}}
\usage{
.default_plf_curve_params()
}
\description{
Example shape parameters used by \code{\link[=plot_plf_curves]{plot_plf_curves()}}
}
\keyword{internal}
