% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_design.R
\name{.normalize_output_files}
\alias{.normalize_output_files}
\title{Normalise \code{output_files} to a named character vector}
\usage{
.normalize_output_files(output_files)
}
\description{
Normalise \code{output_files} to a named character vector
}
\keyword{internal}
