% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_config.R
\docType{data}
\name{.param_expression_ops}
\alias{.param_expression_ops}
\title{Operators allowed in a parameter \code{expression}}
\format{
An object of class \code{character} of length 4.
}
\usage{
.param_expression_ops
}
\description{
Operators allowed in a parameter \code{expression}
}
\keyword{internal}
