% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_config.R
\name{.parse_param_expression}
\alias{.parse_param_expression}
\title{Parse an \code{expression} into a call and the parameter names it references}
\usage{
.parse_param_expression(expression, name, known_names)
}
\description{
Parse an \code{expression} into a call and the parameter names it references
}
\keyword{internal}
