% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_design.R
\name{.read_named_outputs}
\alias{.read_named_outputs}
\title{Read one named set of Daisy output files}
\usage{
.read_named_outputs(output_files, src_root, output_reader, reader_args)
}
\description{
Read one named set of Daisy output files
}
\keyword{internal}
