% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dai_programs.R
\name{.render_program_body}
\alias{.render_program_body}
\title{Shared body (everything but the \verb{(defprogram NAME TYPE-OR-PARENT} header
line and its closing paren) for a base or inheriting program.}
\usage{
.render_program_body(cfg, indent)
}
\description{
Shared body (everything but the \verb{(defprogram NAME TYPE-OR-PARENT} header
line and its closing paren) for a base or inheriting program.
}
\keyword{internal}
