% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_config_scaffold.R
\name{.scaffold_scalar_entry}
\alias{.scaffold_scalar_entry}
\title{Format one scalar \verb{parameters:} entry as YAML text}
\usage{
.scaffold_scalar_entry(name)
}
\description{
Format one scalar \verb{parameters:} entry as YAML text
}
\keyword{internal}
