% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{get_plf_curve}
\alias{get_plf_curve}
\title{Look up a registered custom PLF curve function}
\usage{
get_plf_curve(name, quiet = FALSE)
}
\arguments{
\item{name}{Character scalar.}

\item{quiet}{If \code{TRUE}, return \code{NULL} instead of raising an error when not found.}
}
\value{
The registered function, or \code{NULL}.
}
\description{
Look up a registered custom PLF curve function
}
\keyword{internal}
