% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dai_programs.R
\name{parse_batch_program}
\alias{parse_batch_program}
\title{Inverse of \code{render_batch_program()}.}
\usage{
parse_batch_program(name, children)
}
\description{
Inverse of \code{render_batch_program()}.
}
\keyword{internal}
