% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dai_programs.R
\name{parse_output_entry}
\alias{parse_output_entry}
\title{Inverse of \code{render_output_entry()}.}
\usage{
parse_output_entry(nd)
}
\description{
Inverse of \code{render_output_entry()}.
}
\keyword{internal}
