% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dai_programs.R
\name{parse_program_body}
\alias{parse_program_body}
\title{Inverse of \code{.render_program_body()}.}
\usage{
parse_program_body(children)
}
\description{
Inverse of \code{.render_program_body()}.
}
\keyword{internal}
