% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dai_programs.R
\name{parse_time_field}
\alias{parse_time_field}
\title{Inverse of \code{render_time_field()}.}
\usage{
parse_time_field(children)
}
\description{
Inverse of \code{render_time_field()}.
}
\keyword{internal}
