% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_add_stemID.R
\name{pme_add_stemID}
\alias{pme_add_stemID}
\title{Integration of Segmented Stem Points}
\usage{
pme_add_stemID(las, las_attr)
}
\arguments{
\item{las}{Pointcloud to be reintergrated with the semgented stem points}

\item{las_attr}{Pointcloud of segmented stems from pme_stem_detection_cell, pme_stem_detection_cell_dbscan, or pme_clean_stems}
}
\value{
Returns Pointcloud with stemID as new Variable: las$stemID
}
\description{
Integrates the segmented stem points from pme_stem_detection_cell or pme_stem_detection_cell_dbscan into the original pointcloud
}
