% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_clean_stems.R
\name{pme_clean_stems}
\alias{pme_clean_stems}
\title{Clean Stems}
\usage{
pme_clean_stems(stems, th_nn = 12, th_area = 1.5, th_pts = 100, crs)
}
\arguments{
\item{stems}{Stem product from pme_stem_segm, pme_stem_segm_cell, or pme_stem_segm_dbscan}

\item{th_nn}{Maximum of 50cm height intervals which can be empty, before it is no longer considered a stem}

\item{th_area}{Maximum area of vertically flattened stem points, before it is no longer considered a stem}

\item{th_pts}{Minimum number of points in the stem, before it is no longer considered a stem}

\item{crs}{EPSG-code of CRS in which the data is present}
}
\value{
stems_clean List of sf-dataframe & cleaned pointcloud of stems. Separate with st_as_sf(stems_clean\link{1:5}, coords = c('center_x', 'center_y')) & stems_clean[\link{6}]
}
\description{
Cleans up the segmented stems by pme_stem_segm, pme_stem_segm_cell & pme_stem_segm_dbscan
}
