% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_dbh_estimation.R
\name{pme_dbh_estimation}
\alias{pme_dbh_estimation}
\title{DBH Estimation}
\usage{
pme_dbh_estimation(stems_clean, z_mid, q, crs)
}
\arguments{
\item{stems_clean}{Cleaned stems from pme_clean_stems}

\item{z_mid}{Height of middle of DBH extraction heigh}

\item{q}{Quantile of distances between all flattened stem points}

\item{crs}{EPSG-code of CRS in which the data is present}
}
\value{
sf-dataframe. Spatial dataframe with the estimated DBHs
}
\description{
Estimates the DBH vales based on segmented and cleaned stems.
}
