% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_intfix.R
\name{pme_intfix}
\alias{pme_intfix}
\title{GS-100C+ Intensity with flightlines}
\usage{
pme_intfix(input_path, drone_path, crs)
}
\arguments{
\item{input_path}{Path to LAS file}

\item{drone_path}{Path to GNSS flight file}

\item{crs}{EPSG code of the LiDAR data}
}
\value{
NULL Writes the corrected LAS File into the input directory
}
\description{
Corrects the flight lines effect from raw, stitched GS-100C+ LiDAR data
}
