% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_stem_detection_cell.R
\name{pme_stem_detection_cell}
\alias{pme_stem_detection_cell}
\title{Stem detection algorithm based on DBSCAN & HDBSCAN}
\usage{
pme_stem_detection_cell(
  las,
  minPts = 100,
  cellsize = 5,
  hmax = 10,
  hmin = 0.5,
  minPts_post = 150,
  eps = 0.12
)
}
\arguments{
\item{las}{Pointcloud data to be segmented}

\item{minPts}{Minimum points for the HDBSCAN on initial clustering to be considered as a stem. Higher values disregards more less dense clusters}

\item{cellsize}{Cell size of initial clustering. Higher cell sized will require considerably more RAM! CAUTION!!!}

\item{hmax}{Maximum height of point cloud. Points above are disregarded in segmentation}

\item{hmin}{Minimum height of point cloud. Points below are disregarded in segmentation}

\item{minPts_post}{Minimum points for the DBSCAN on secondary clustering to be considered as a stem. Higher values disregards more less dense clusters}

\item{eps}{Epsilon of HDBSCAN/DBSCAN. Maximum distance between two points to be considered part of the stem. Smaller values segment stricter, but might remove some stem points}
}
\description{
Segments stems by cells in height-normalised LiDAR data, based on HDBSCAN, then DBSCAN
}
