% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_stem_detection_cell_dbscan.R
\name{pme_stem_detection_cell_dbscan}
\alias{pme_stem_detection_cell_dbscan}
\title{Stem detection algorithm based on DBSCAN & only}
\usage{
pme_stem_detection_cell_dbscan(
  las,
  minPts = 100,
  cellsize = 25,
  hmax = 12,
  hmin = 0.5,
  minPts_post = 125,
  eps = 0.12
)
}
\arguments{
\item{las}{Pointcloud data to be segmented}

\item{minPts}{Minimum points for the HDBSCAN on initial clustering to be considered as a stem. Higher values disregards more less dense clusters}

\item{cellsize}{Cell size of initial clustering. Higher cell sized will require considerably more RAM! CAUTION!!!}

\item{hmax}{Maximum height of point cloud. Points above are disregarded in segmentation}

\item{hmin}{Minimum height of point cloud. Points below are disregarded in segmentation}

\item{minPts_post}{Minimum points for the DBSCAN on secondary clustering to be considered as a stem. Higher values disregards more less dense clusters}

\item{eps}{Epsilon of DBSCAN. Maximum distance between two points to be considered part of the stem. Smaller values segment stricter, but might remove some stem points}
}
\description{
Segments stems by cells in height-normalised LiDAR data, based on two step DBSCAN clustering, works singificantly faster then pme_stem_detection_cell
}
