% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_tree_curvature.R
\name{pme_tree_curvature}
\alias{pme_tree_curvature}
\title{Stem Curvature Estimation}
\usage{
pme_tree_curvature(stems, chm = chm, crs)
}
\arguments{
\item{stems}{Cleaned stem product from pme_clean_stems}

\item{chm}{Canopy height model which overlaps all stems}

\item{crs}{EPSG-code of CRS in which the data is present}
}
\description{
Estimates the stem curvature and tree height from the cleaned and segmented stems from pme_clean_stems
}
