% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_tree_curvature_plot.R
\name{pme_tree_curvature_plot_check}
\alias{pme_tree_curvature_plot_check}
\title{Tree Curvature Plot}
\usage{
pme_tree_curvature_plot_check(stems_clean, stemID)
}
\arguments{
\item{stems_clean}{Cleaned stems from pme_clean_stems}

\item{stemID}{Stem ID of the stem to plot}
}
\description{
Creates a visual plot of a single stems curvature estimation
}
