% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pme_understory_detection.R
\name{pme_understory_detection}
\alias{pme_understory_detection}
\title{Understory Segmentation}
\usage{
pme_understory_detection(las, res = 25, bp_pre = 20, min_uth = 5)
}
\arguments{
\item{stems}{Cleaned stem product from pme_clean_stems}

\item{chm}{Canopy height model which overlaps all stems}

\item{crs}{EPSG-code of CRS in which the data is present}
}
\description{
Segments the pointcloud from understory segmentation based on pme_stem_detection_cell or pme_stem_detection_cell_dbscan
}
