% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dai_columns.R
\name{render_column}
\alias{render_column}
\title{Render one \code{columns[[i]]} entry to a complete \verb{(defcolumn ...)} block.}
\usage{
render_column(cfg)
}
\description{
Render one \code{columns[[i]]} entry to a complete \verb{(defcolumn ...)} block.
}
\keyword{internal}
