% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_config.R
\name{param_config_default_values}
\alias{param_config_default_values}
\title{Default values for every config parameter}
\usage{
param_config_default_values(config)
}
\arguments{
\item{config}{A \code{daisyr_param_config} object.}
}
\value{
Named numeric vector, one value per \code{\link[=param_config_names]{param_config_names()}}.
}
\description{
Default values for every config parameter
}
